+++ BookManagement/src/main/java/services/BookService.java
@@ -3,6 +3,8 @@ package services;
 import dao.AuthorDao;
 import dao.BookDao;
 import entities.BookModel;
+import exceptions.DuplicateException;
+import exceptions.EntityNotFoundException;
 
 import java.util.List;
 
@@ -15,8 +17,13 @@ public class BookService {
         this.bookDao = bookDao;
     }
 
-    public void addBook(BookModel newBook){
+    public void addBook(BookModel newBook) throws DuplicateException {
         List<BookModel> books = bookDao.getAllBooks();
+        for (BookModel book : books){
+            if ((book.getTitle()).equals((newBook.getTitle()))){
+                throw new DuplicateException();
+            }
+        }
         bookDao.addBook(newBook);
     }
     public List<BookModel> viewBooks(){
@@ -24,8 +31,11 @@ public class BookService {
         return books;
     }
 
-    public void deleteBook(int id){
+    public void deleteBook(int id) throws EntityNotFoundException {
         BookModel book = bookDao.findById(id);
+        if (book == null){
+            throw new EntityNotFoundException();
+        }
         bookDao.deleteBook(book);
     }
